Implement ExternalAuthTrait and ExternalAuthInterface
So other backends can support external auth.
