+++ library/Icinga/Authentication/User/ExternalBackend.php
@@ -10,14 +10,9 @@ use Icinga\User;
 /**
  * Test login with external authentication mechanism, e.g. Apache
  */
-class ExternalBackend implements UserBackendInterface
+class ExternalBackend implements UserBackendInterface, ExternalAuthInterface
 {
-    /**
-     * Possible variables where to read the user from
-     *
-     * @var string[]
-     */
-    public static $remoteUserEnvvars = array('REMOTE_USER', 'REDIRECT_REMOTE_USER');
+    use ExternalAuthTrait;
 
     /**
      * The name of this backend
@@ -61,64 +56,25 @@ class ExternalBackend implements UserBackendInterface
     }
 
     /**
-     * Get the remote user from environment or $_SERVER, if any
-     *
-     * @param   string  $variable   The name of the variable where to read the user from
-     *
-     * @return  string|null
-     */
-    public static function getRemoteUser($variable = 'REMOTE_USER')
-    {
-        $username = getenv($variable);
-        if ($username !== false) {
-            return $username;
-        }
-
-        if (array_key_exists($variable, $_SERVER)) {
-            return $_SERVER[$variable];
-        }
-    }
-
-    /**
-     * Get the remote user information from environment or $_SERVER, if any
-     *
-     * @return  array   Contains always two entries, the username and origin which may both set to null.
-     */
-    public static function getRemoteUserInformation()
-    {
-        foreach (static::$remoteUserEnvvars as $envVar) {
-            $username = static::getRemoteUser($envVar);
-            if ($username !== null) {
-                return array($username, $envVar);
-            }
-        }
-
-        return array(null, null);
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function authenticate(User $user, $password = null)
     {
-        list($username, $field) = static::getRemoteUserInformation();
-        if ($username !== null) {
-            $user->setExternalUserInformation($username, $field);
-
-            if ($this->stripUsernameRegexp) {
-                $stripped = @preg_replace($this->stripUsernameRegexp, '', $username);
-                if ($stripped === false) {
-                    Logger::error('Failed to strip external username. The configured regular expression is invalid.');
-                    return false;
-                }
-
-                $username = $stripped;
-            }
-
-            $user->setUsername($username);
-            return true;
+        $external = $this->authenticateExternal($user);
+        if (! $external) {
+            return false;
         }
 
-        return false;
+        if ($this->stripUsernameRegexp) {
+            $stripped = @preg_replace($this->stripUsernameRegexp, '', $external[0]);
+            if ($stripped === false) {
+                Logger::error('Failed to strip external username. The configured regular expression is invalid.');
+                return false;
+            }
+
+            $user->setUsername($stripped);
+        }
+
+        return true;
     }
 }
